[100] The 3n + 1 problem
This problem is solvable using straight forward method.
However, in order to speed up the program, I use some tricks:
The first one is using a table to store calculated LENGTH to prevent
recalculations.
The second one is dividing all possible input into regions,
and precaculate the max LENGTHs of the regions. So when input i, j
are across multiple regions, the program can lookup quickly.